removed /ssw base path
